refactor(ledger): collocate the operation route registrars :hammer:
The Huma terminal registrar came out of operation_handler.go and the Fiber
guard-chain registrars came out of routes.go; both now live in
operation_routes.go.

The moved Huma block references the postgres operation adapter for its typed
PATCH body, so operation_routes.go imports it alongside the pkgHTTP-aliased
midaz net/http.

No route, authz tuple, operation ID or path changed.
