+++ components/ledger/internal/adapters/http/in/operation_handler.go
@@ -204,51 +204,3 @@ func (handler *OperationHandler) UpdateOperation(ctx context.Context, in *Update
 
 	return &UpdateOperationResponse{Status: http.StatusOK, Body: op}, nil
 }
-
-// RegisterOperationRoutes registers the two operation read ops plus the PATCH
-// (money-write leg) on the shared Huma API. It is the per-file seam the unified
-// server calls; the auth (auth.Authorize("midaz","operations",verb)) + tenant +
-// ParseUUIDPathParameters("operation") chain for these routes is attached in the unified
-// server (Fiber level) BEFORE the Huma terminal, not here. Paths are GROUP-RELATIVE (the
-// group's PrefixModifier writes the version into each op's op.Path, not into a servers entry).
-//
-// opSuffix distinguishes the operation IDs one version group publishes from another's — see
-// routeOpSuffixV1. The v1 group passes the empty suffix so its IDs stay exactly what published
-// SDKs bind to; the v2 group passes "V2" so its twins do not collide in the one document.
-func RegisterOperationRoutes(api huma.API, h *OperationHandler, opSuffix string) {
-	const (
-		listPath  = "/organizations/{organization_id}/ledgers/{ledger_id}/accounts/{account_id}/operations"
-		idPath    = listPath + "/{operation_id}"
-		patchPath = "/organizations/{organization_id}/ledgers/{ledger_id}/transactions/{transaction_id}/operations/{operation_id}"
-		tag       = "Operations"
-	)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "getAllOperationsByAccount" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        listPath,
-		Summary:     "Get all Operations by account",
-		Tags:        []string{tag},
-		Security:    secOperationBearer,
-	}, h.GetAllOperationsByAccount)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "getOperationByAccount" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        idPath,
-		Summary:     "Get Operation",
-		Tags:        []string{tag},
-		Security:    secOperationBearer,
-	}, h.GetOperationByAccount)
-
-	huma.Register(api, huma.Operation{
-		OperationID:      "updateOperation" + opSuffix,
-		Method:           http.MethodPatch,
-		Path:             patchPath,
-		Summary:          "Update an Operation",
-		Tags:             []string{tag},
-		Security:         secTransactionBearer, // BearerAuth (Bearer-only), matching the Fiber guard chain on the transaction-path PATCH.
-		SkipValidateBody: true,                 // body validated imperatively (http.DecodeAndValidate) — plain decode, not merge-patch.
-	}, h.UpdateOperation)
-	attachTypedRequestBody[operation.UpdateOperationInput](api, "updateOperation"+opSuffix)
-}
